requirements.txt
# Production dependencies
fastapi==0.116.0
uvicorn==0.24.0
boto3==1.34.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
redis==4.5.0
typing-extensions>=4.10.0
python-multipart==0.0.6
gunicorn==21.2.0 
pinecone-client==6.0.0
beautifulsoup4==4.12.2
slowapi==0.1.9
numpy==1.26.4
flask==3.0.0

# Development dependencies
ruff==0.4.8
black==24.4.2
isort==5.13.2
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.2